+++ src/main/java/net/ddraig/suprememiningdimension/procedures/StrangeForestsPlayerEntersDimensionProcedure.java
@@ -31,29 +31,30 @@ public class StrangeForestsPlayerEntersDimensionProcedure {
 				xx = entity.getX();
 				zz = entity.getZ();
 				yy = world.getHeight(Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, (int) xx, (int) zz);
-				if (world.canSeeSkyFromBelowWater(new BlockPos(xx, yy, zz))) {
+				if (world.canSeeSkyFromBelowWater(new BlockPos((int)xx, (int)yy, (int)zz))) {
 					{
 						Entity _ent = entity;
 						_ent.teleportTo(xx, yy, zz);
 						if (_ent instanceof ServerPlayer _serverPlayer)
 							_serverPlayer.connection.teleport(xx, yy, zz, _ent.getYRot(), _ent.getXRot());
 					}
-					world.setBlock(new BlockPos(xx - 0, yy - 1, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
-					world.setBlock(new BlockPos(xx - 1, yy - 1, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
-					world.setBlock(new BlockPos(xx - 2, yy - 0, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
-					world.setBlock(new BlockPos(xx + 1, yy - 0, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
-					world.setBlock(new BlockPos(xx - 2, yy + 1, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
-					world.setBlock(new BlockPos(xx + 1, yy + 1, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
-					world.setBlock(new BlockPos(xx - 2, yy + 2, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
-					world.setBlock(new BlockPos(xx + 1, yy + 2, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
-					world.setBlock(new BlockPos(xx - 0, yy + 3, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
-					world.setBlock(new BlockPos(xx - 1, yy - 3, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
-					world.setBlock(new BlockPos(xx - 2, yy - 3, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
-					world.setBlock(new BlockPos(xx + 1, yy + 3, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
-					world.setBlock(new BlockPos(xx + 1, yy - 1, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
-					world.setBlock(new BlockPos(xx - 2, yy - 1, zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx - 0), (int)(yy - 1), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx - 1), (int)(yy - 1), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx - 2), (int)(yy - 0), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx + 1), (int)(yy - 0), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx - 2), (int)(yy + 1), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx + 1), (int)(yy + 1), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx - 2), (int)(yy + 2), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx + 1), (int)(yy + 2), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx - 0), (int)(yy + 3), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx - 1), (int)(yy - 3), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx - 2), (int)(yy - 3), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx + 1), (int)(yy + 3), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx + 1), (int)(yy - 1), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+						world.setBlock(new BlockPos((int)(xx - 2), (int)(yy - 1), (int)zz), SupremeMiningDimensionModBlocks.ETCHED_WOOD.get().defaultBlockState(), 3);
+
 					if (world instanceof Level _level)
-						StrangeForestsPortalBlock.portalSpawn(_level, new BlockPos(xx, yy, zz));
+						StrangeForestsPortalBlock.portalSpawn(_level, new BlockPos((int)xx, (int)yy, (int)zz));
 				}
 			}
 		}
